requirements-test.txt
ansible-core>=2.16,<2.19
ansible-lint>=24.12.0
pytest>=8.0.0
PyYAML>=6.0.1
ruamel.yaml>=0.18.11
yamllint>=1.35.1
